backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openheart-backend"
version = "0.1.0"
description = "OpenHeart Cyprus - Cardiology EMR Backend"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "OpenHeart Cyprus", email = "dev@openheart.cy"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]

dependencies = [
    # FastAPI & Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "pgvector>=0.2.4",

    # Redis
    "redis>=5.0.1",

    # Validation & Settings
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",
    "pyotp>=2.9.0",

    # HTTP Client
    "httpx>=0.26.0",

    # File Processing
    "python-magic>=0.4.27",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "pillow>=10.2.0",

    # FHIR
    "fhir.resources>=7.1.0",

    # HL7 v2
    "hl7apy>=1.3.4",

    # S3/MinIO
    "boto3>=1.34.0",

    # Utilities
    "python-dateutil>=2.8.2",
    "orjson>=3.9.12",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",

    # Code Quality
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",

    # Security Scanning
    "bandit>=1.7.7",

    # Type Stubs
    "types-python-dateutil>=2.8.19",
    "types-passlib>=1.7.7",
    "types-redis>=4.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | migrations
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "app/tests"]
addopts = "-v --cov=app --cov-report=term-missing"

[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B101"]  # Skip assert warnings in tests
